Change configid_intptr_* to configid_handle_* with uint64_t as its type + related adjustments
- Change configid_intptr_* to configid_handle_* with uint64_t as its type
- Fix missing handling for global intptr/handle config values
- Remove x86/32-bit build configuration, since config_handle_* values cannot be used in window messages in 32-bit builds
- Remove line in README.md mentioning 32-bit support
- Streamline float window message config value handling

+++ src/DesktopPlusUI/TextureManager.h
@@ -96,7 +96,7 @@ class TextureManager
         bool GetTextureInfo(const CustomAction& action, ImVec2& size, ImVec2& uv_min, ImVec2& uv_max) const;
 
         int  GetWindowIconCacheID(HWND window_handle); //Returns -1 on error
-        int  GetWindowIconCacheID(HWND window_handle, intptr_t& icon_handle_config); //Updates icon_handle_config when lookup with window_handle succeeds or falls back to icon_handle_config
+        int  GetWindowIconCacheID(HWND window_handle, uint64_t& icon_handle_config); //Updates icon_handle_config when lookup with window_handle succeeds or falls back to icon_handle_config
         int  GetWindowIconCacheID(HICON icon_handle);  //Returns -1 on error
         bool GetWindowIconTextureInfo(int icon_cache_id, ImVec2& size, ImVec2& uv_min, ImVec2& uv_max) const;
